requirements.txt
PyYAML~=6.0
numpy~=1.19.5
pandas~=1.1.5
scikit-learn~=0.22.2.post1
torch~=1.9.0
transformers~=4.12.2
wandb~=0.12.6
slugify~=0.0.1
scipy~=1.7.1